+++ app/src/main/java/com/example/ezchat/activities/LoginPhoneNumberActivity.java
@@ -1,4 +1,4 @@
-package com.example.ezchat;
+package com.example.ezchat.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.ezchat.R;
 import com.hbb20.CountryCodePicker;
 
 public class LoginPhoneNumberActivity extends AppCompatActivity {
