+++ app/src/main/java/com/example/ezchat/activities/MainActivity.java
@@ -1,14 +1,16 @@
-package com.example.ezchat;
+package com.example.ezchat.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.MenuItem;
 import android.widget.ImageButton;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.ezchat.ChatFragment;
+import com.example.ezchat.ProfileFragment;
+import com.example.ezchat.R;
 import com.example.ezchat.utils.FirebaseUtil;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
