--- app/src/main/res/drawable/btn_rounded_corner.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape android:shape="rectangle" xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <stroke android:width="1dp" android:color="@color/white"/>
-    <corners android:radius="40dp"/>
-    <solid android:color="@color/white"/>
-</shape>
