+++ app/src/main/res/layout/activity_chat.xml
@@ -66,7 +66,7 @@
             android:layout_marginEnd="10dp"
             android:layout_marginBottom="10dp"
             android:layout_toLeftOf="@id/message_send_btn"
-            android:background="@drawable/edit_text_rounded_corner"
+            android:background="@drawable/shape_rectangle"
             android:elevation="4dp"
             android:hint="@string/write_message_here"
             android:padding="10dp"
