+++ app/src/main/res/layout/activity_login_user_name.xml
@@ -9,7 +9,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:weightSum="100"
-    tools:context=".LoginUserNameActivity">
+    tools:context=".activities.LoginUserNameActivity">
 
 
     <TextView
@@ -30,7 +30,7 @@
             android:layout_width="32dp"
             android:layout_height="32dp"
             android:layout_margin="10dp"
-            android:background="@drawable/circular_bg"
+            android:background="@drawable/shape_circle"
             android:backgroundTint="@color/light_gray">
 
             <TextView
@@ -45,7 +45,7 @@
             android:layout_width="32dp"
             android:layout_height="32dp"
             android:layout_margin="10dp"
-            android:background="@drawable/circular_bg"
+            android:background="@drawable/shape_circle"
             android:backgroundTint="@color/light_gray">
             <TextView
                 android:layout_width="wrap_content"
@@ -59,7 +59,7 @@
             android:layout_width="32dp"
             android:layout_height="32dp"
             android:layout_margin="10dp"
-            android:background="@drawable/circular_bg"
+            android:background="@drawable/shape_circle"
             android:backgroundTint="@color/my_primary">
             <TextView
                 android:layout_width="wrap_content"
@@ -75,7 +75,7 @@
         android:layout_width="110dp"
         android:layout_height="90dp"
         android:layout_weight="20"
-        android:src="@drawable/person_icon"
+        android:src="@drawable/icon_person"
         app:tint="@color/my_primary"
         />
 
@@ -101,7 +101,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
-            android:background="@drawable/edit_text_rounded_corner"
+            android:background="@drawable/shape_rectangle"
             android:elevation="5dp"
             android:hint="Username"
             android:inputType="text"
@@ -113,7 +113,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
-            android:background="@drawable/edit_text_rounded_corner"
+            android:background="@drawable/shape_rectangle"
             android:elevation="5dp"
             android:hint="Password"
             android:inputType="text"
@@ -125,7 +125,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
-            android:background="@drawable/edit_text_rounded_corner"
+            android:background="@drawable/shape_rectangle"
             android:elevation="5dp"
             android:hint="Confirm Password"
             android:inputType="text"
@@ -145,7 +145,7 @@
             android:id="@+id/login_code_btn"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:background="@drawable/btn_rounded_corner"
+            android:background="@drawable/shape_rectangle"
             android:text=" SIGN UP " />
 
         <ProgressBar
