--- app/src/main/res/layout/chat_message_recycler_row.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@drawable/edit_text_rounded_corner"
-        android:backgroundTint="@color/chat_color_receiver"
-        android:padding="8dp"
-        android:layout_marginEnd="80dp"
-        android:layout_marginBottom="8dp"
-        android:id="@+id/left_chat_layout">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/white"
-        android:textSize="18sp"
-        android:id="@+id/left_chat_textview"/>
-</LinearLayout>
-
-<LinearLayout
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:background="@drawable/edit_text_rounded_corner"
-    android:backgroundTint="@color/chat_color_sender"
-    android:layout_alignParentEnd="true"
-    android:padding="8dp"
-    android:layout_marginStart="80dp"
-    android:layout_marginBottom="8dp"
-    android:id="@+id/right_chat_layout">
-
-<TextView
-android:layout_width="wrap_content"
-android:layout_height="wrap_content"
-android:textColor="@color/white"
-android:textSize="18sp"
-android:id="@+id/right_chat_textview"/>
-    </LinearLayout>
-
-    </RelativeLayout>
