--- app/src/main/res/layout/fragment_calendar.xml
@@ -1,56 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="16dp">
-
-    <!-- Back Button -->
-    <Button
-        android:id="@+id/back_to_profile_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="start"
-        android:layout_marginBottom="20dp"
-        android:background="@drawable/btn_rounded_corner"
-        android:backgroundTint="@color/info"
-        android:text="Back"
-        android:textColor="@color/white" />
-
-    <!-- Native CalendarView -->
-    <CalendarView
-        android:id="@+id/calendar_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="16dp" />
-
-    <!-- Tasks Header -->
-    <TextView
-        android:id="@+id/task_header"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="8dp"
-        android:text="Tasks for Selected Date"
-        android:textSize="16sp"
-        android:textColor="@color/black"
-        android:textStyle="bold" />
-
-    <!-- RecyclerView for Task List -->
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/task_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:padding="8dp"
-        android:scrollbars="vertical" />
-
-    <!-- Add Task Button -->
-    <Button
-        android:id="@+id/add_task_btn"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:backgroundTint="@color/black"
-        android:background="@drawable/btn_rounded_corner"
-        android:text="Add Task"
-        android:textColor="@color/white" />
-</LinearLayout>
