+++ app/src/main/res/layout/profile_pic_view.xml
@@ -7,8 +7,8 @@
     <ImageView
         android:layout_width="52dp"
         android:layout_height="52dp"
-        android:background="@drawable/circular_bg"
-        android:src="@drawable/person_icon"
+        android:background="@drawable/shape_circle"
+        android:src="@drawable/icon_person"
         android:padding="4dp"
         app:tint="@color/light_gray"
         android:backgroundTint="@color/off_white"
