--- app/src/main/res/layout/recent_chat_recycler_row.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:background="@drawable/edit_text_rounded_corner"
-    android:layout_margin="10dp"
-    android:elevation="2dp">
-
-    <include layout="@layout/profile_pic_view"
-        android:layout_width="52dp"
-        android:layout_height="52dp"
-        android:layout_gravity="center"
-        android:layout_margin="10dp"
-        />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_gravity="center">
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                tools:text="User name"
-                android:textSize="18sp"
-                android:textStyle="bold"
-                android:id="@+id/user_name_text"/>
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                tools:text="10:00"
-                android:layout_alignParentEnd="true"
-                android:layout_marginEnd="10dp"
-                android:id="@+id/last_message_time_text"/>
-        </RelativeLayout>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            tools:text="+000000000000"
-            android:id="@+id/last_message_text"/>
-    </LinearLayout>
-</LinearLayout>
