--- app/src/main/res/layout/search_user_recycler_row.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@drawable/edit_text_rounded_corner"
-    android:layout_margin="10dp"
-    android:elevation="2dp">
-
-    <include layout = "@layout/profile_pic_view"
-    android:layout_width="52dp"
-    android:layout_height="52dp"
-    android:layout_gravity="center"
-    android:layout_margin="10dp"
-    />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_gravity="center">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/user_name"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            android:id="@+id/user_name_text"/>
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="+0(000)-000-0000"
-            android:id="@+id/phone_text"/>
-
-    </LinearLayout>
-
-</LinearLayout>
-
