+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -2,12 +2,12 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
         android:id="@+id/menu_profile"
-        android:icon="@drawable/person_icon"
+        android:icon="@drawable/icon_person"
         android:title="Profile"
         />
     <item
         android:id="@+id/menu_chat"
-        android:icon="@drawable/chat_icon"
+        android:icon="@drawable/icon_chat"
         android:title="Chats"
         />
 </menu>
